separate tests and add tests for enums at endpoints
