+++ src/lexer-models.test.ts
@@ -3,7 +3,6 @@ import { Lexer, isPunctuatorTokenKind } from './lexer';
 import { Token, TokenKind } from './token';
 import { inspect } from 'util';
 import { dedent } from './__testsUtils__/dedent';
-import { AnySpecError } from './error/AnySpecError';
 
 function lexFirst(str: string) {
   const lexer = new Lexer(new Source(str));
@@ -16,7 +15,7 @@ function lexSecond(str: string) {
   return lexer.advance();
 }
 
-const getFullTokenList = (source: Source) => {
+export const getFullTokenList = (source: Source) => {
   const lexer = new Lexer(source);
 
   let tokenList = [];
@@ -32,6 +31,7 @@ const getFullTokenList = (source: Source) => {
 
   return tokenList;
 };
+
 
 function expectSyntaxErrorFirst(text: string, expectedError: any) {
   let error = null;
@@ -719,349 +719,6 @@ describe('lexer can catch some parenthesis errors inside enums', () => {
 
     expect(() => getFullTokenList(enumString)).toThrow(
       'Syntax Error: parenthesis should be balanced inside enum definition',
-    );
-  });
-});
-
-describe('lexer can understand endpoints', () => {
-  it('lexer can understand endpoints basic', () => {
-    const sourceString = `
-\`/analytics_events\`:
-    // **Create**
-    @token POST /analytics_events AnalyticsEventNewRequest
-      => AnalyticsEventNewResponse
-    POST /analytics_events AnalyticsEventNewRequest
-      => AnalyticsEventNewResponse
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    const expectedTokens = [
-      '/analytics_events',
-      ':',
-      '<INDENT>',
-      'Description',
-      '@',
-      'token',
-      'POST',
-      '/analytics_events',
-      'AnalyticsEventNewRequest',
-      '<INDENT>',
-      '=>',
-      'AnalyticsEventNewResponse',
-      '<DEDENT>',
-      'POST',
-      '/analytics_events',
-      'AnalyticsEventNewRequest',
-      '<INDENT>',
-      '=>',
-      'AnalyticsEventNewResponse',
-      '<DEDENT>',
-      '<DEDENT>',
-    ];
-
-    expect(getFullTokenList(enumString)).toEqual(expectedTokens);
-  });
-
-  it('lexer can understand endpoints basic, should ignore empty lines', () => {
-    const sourceString = `
-\`/analytics_events\`:
-  // **Create**
-  @token POST /analytics_events AnalyticsEventNewRequest
-    => AnalyticsEventNewResponse
-              
-
-
-  POST /analytics_events AnalyticsEventNewRequest
-    => AnalyticsEventNewResponse
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    const expectedTokens = [
-      '/analytics_events',
-      ':',
-      '<INDENT>',
-      'Description',
-      '@',
-      'token',
-      'POST',
-      '/analytics_events',
-      'AnalyticsEventNewRequest',
-      '<INDENT>',
-      '=>',
-      'AnalyticsEventNewResponse',
-      '<DEDENT>',
-      'POST',
-      '/analytics_events',
-      'AnalyticsEventNewRequest',
-      '<INDENT>',
-      '=>',
-      'AnalyticsEventNewResponse',
-      '<DEDENT>',
-      '<DEDENT>',
-    ];
-
-    expect(getFullTokenList(enumString)).toEqual(expectedTokens);
-  });
-
-  it('lexer can understand endpoints basic v2', () => {
-    const sourceString = `
-\`/analytics_events\`:
-  // **Create**
-  @token POST /analytics_events AnalyticsEventNewRequest
-    => AnalyticsEventNewResponse
-    
-POST /analytics_events AnalyticsEventNewRequest
-  => AnalyticsEventNewResponse
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    const expectedTokens = [
-      '/analytics_events',
-      ':',
-      '<INDENT>',
-      'Description',
-      '@',
-      'token',
-      'POST',
-      '/analytics_events',
-      'AnalyticsEventNewRequest',
-      '<INDENT>',
-      '=>',
-      'AnalyticsEventNewResponse',
-      '<DEDENT>',
-      '<DEDENT>',
-      'POST',
-      '/analytics_events',
-      'AnalyticsEventNewRequest',
-      '<INDENT>',
-      '=>',
-      'AnalyticsEventNewResponse',
-      '<DEDENT>',
-    ];
-
-    expect(getFullTokenList(enumString)).toEqual(expectedTokens);
-  });
-
-  it('lexer can understand endpoints basic v3', () => {
-    const sourceString = `
-GET /hero/exchangeRates?HrGetExchangeRateRequestQuery
-  => HrGetExchangeRateListResponse
-  
-  
-\`/roles\`:
-  @token $CRUDL /roles
-  
-HEAD /pechkin/mandrill/event
-  => {}
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    const expectedTokens = [
-      'GET',
-      '/hero/exchangeRates?HrGetExchangeRateRequestQuery',
-      '<INDENT>',
-      '=>',
-      'HrGetExchangeRateListResponse',
-      '<DEDENT>',
-      '/roles',
-      ':',
-      '<INDENT>',
-      '@',
-      'token',
-      '$',
-      'CRUDL',
-      '/roles',
-      '<DEDENT>',
-      'HEAD',
-      '/pechkin/mandrill/event',
-      '<INDENT>',
-      '=>',
-      '{',
-      '}',
-      '<DEDENT>',
-    ];
-
-    expect(getFullTokenList(enumString)).toEqual(expectedTokens);
-  });
-  it('lexer can understand endpoints basic v4', () => {
-    const sourceString = `
-
-HEAD /pechkin/mandrill/event
-    => {}
-  
-// **Delete**
-      // Permissions: \`companies:write\`
-      @token DELETE /accounting/bank_accounts/:id:i
-          => 204
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    const expectedTokens = [
-      'HEAD',
-      '/pechkin/mandrill/event',
-      '<INDENT>',
-      '=>',
-      '{',
-      '}',
-      '<DEDENT>',
-      'Description',
-      '<INDENT>',
-      'Description',
-      '@',
-      'token',
-      'DELETE',
-      '/accounting/bank_accounts/:id:i',
-      '<INDENT>',
-      '=>',
-      '204',
-      '<DEDENT>',
-      '<DEDENT>',
-    ];
-
-    expect(getFullTokenList(enumString)).toEqual(expectedTokens);
-  });
-  it('lexer can understand endpoints complex', () => {
-    const sourceString = `
-
-HEAD /pechkin/mandrill/event {messageSenderId?: i, conversationId?: i, ticketId?: i, taskId?: i, isNewTicketRequired?: b }
-    => {tigerDocument: TigerDocument, pandaDocument: PandaDocument, messageIds: i[]}
-  
-// **Delete**
-      // Permissions: \`companies:write\`
-      @token DELETE /accounting/bank_accounts/:id:i
-          => 204
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    const expectedTokens = [
-      'HEAD',
-      '/pechkin/mandrill/event',
-      '{',
-      'messageSenderId',
-      '?',
-      ':',
-      'i',
-      'conversationId',
-      '?',
-      ':',
-      'i',
-      'ticketId',
-      '?',
-      ':',
-      'i',
-      'taskId',
-      '?',
-      ':',
-      'i',
-      'isNewTicketRequired',
-      '?',
-      ':',
-      'b',
-      '}',
-      '<INDENT>',
-      '=>',
-      '{',
-      'tigerDocument',
-      ':',
-      'TigerDocument',
-      'pandaDocument',
-      ':',
-      'PandaDocument',
-      'messageIds',
-      ':',
-      'i',
-      '[',
-      ']',
-      '}',
-      '<DEDENT>',
-      'Description',
-      '<INDENT>',
-      'Description',
-      '@',
-      'token',
-      'DELETE',
-      '/accounting/bank_accounts/:id:i',
-      '<INDENT>',
-      '=>',
-      '204',
-      '<DEDENT>',
-      '<DEDENT>',
-    ];
-
-    expect(getFullTokenList(enumString)).toEqual(expectedTokens);
-  });
-});
-
-describe('lexer can catch errors in endpoints', () => {
-  it('unbalanced indentation not allowed', () => {
-    const sourceString = `
-  \`/analytics_events\`:
-  // **Create**
- @token POST /analytics_events AnalyticsEventNewRequest
-    => AnalyticsEventNewResponse
-     POST /analytics_events AnalyticsEventNewRequest
-    => AnalyticsEventNewResponse
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    expect(() => getFullTokenList(enumString)).toThrow(
-      'Syntax Error: unindent does not match any outer indentation level',
-    );
-  });
-  it('unbalanced indentation not allowed 1', () => {
-    const sourceString = `
-  \`/analytics_events\`:
-  // **Create**
-  @token POST /analytics_events AnalyticsEventNewRequest
-    => AnalyticsEventNewResponse
-
-   POST /analytics_events AnalyticsEventNewRequest
-    => AnalyticsEventNewResponse
-`;
-    const enumString = new Source(
-      sourceString,
-      'Tinyspec endpoints code',
-      { line: 1, column: 1 },
-      'endpoints',
-    );
-
-    expect(() => getFullTokenList(enumString)).toThrow(
-      'Syntax Error: unindent does not match any outer indentation level',
     );
   });
 });
